Controller refactor (#789)
* refactor controllers, reduce repetition

* Use inheritance to limit repetition in controllers
