+++ app/controllers/hyrax/masters_papers_controller.rb
@@ -2,10 +2,7 @@
 #  `rails generate hyrax:work MastersPaper`
 
 module Hyrax
-  class MastersPapersController < ApplicationController
-    # Adds Hyrax behaviors to the controller.
-    include Hyrax::WorksControllerBehavior
-    include Hyrax::BreadcrumbsForWorks
+  class MastersPapersController < HycController
     self.curation_concern_type = ::MastersPaper
 
     # Use this line if you want to use a custom presenter
@@ -13,14 +10,5 @@ module Hyrax
 
     before_action :ensure_admin!, only: :destroy
     before_action :ensure_admin_set!, only: [:create, :new, :edit, :update]
-
-    private
-    def ensure_admin!
-      authorize! :read, :admin_dashboard
-    end
-
-    def ensure_admin_set!
-      return redirect_to root_path, alert: 'No Admin Sets have been created.' if AdminSet.all.count.zero?
-    end
   end
 end
